+++ src/js/core/cover.js
@@ -26,25 +26,38 @@ export default {
 
     update: {
         read() {
-            const el = this.$el;
-            const { offsetHeight: height, offsetWidth: width } =
-                getPositionedParent(el) || parent(el);
-            const dim = Dimensions.cover(
-                {
-                    width: this.width || el.naturalWidth || el.videoWidth || el.clientWidth,
-                    height: this.height || el.naturalHeight || el.videoHeight || el.clientHeight,
-                },
-                {
-                    width: width + (width % 2 ? 1 : 0),
-                    height: height + (height % 2 ? 1 : 0),
-                }
-            );
+            const { ratio, cover } = Dimensions;
+            const { $el, width, height } = this;
+
+            let dim = { width, height };
 
             if (!dim.width || !dim.height) {
+                const intrinsic = {
+                    width: $el.naturalWidth || $el.videoWidth || $el.clientWidth,
+                    height: $el.naturalHeight || $el.videoHeight || $el.clientHeight,
+                };
+
+                if (dim.width) {
+                    dim = ratio(intrinsic, 'width', dim.width);
+                } else if (height) {
+                    dim = ratio(intrinsic, 'height', dim.height);
+                } else {
+                    dim = intrinsic;
+                }
+            }
+
+            const { offsetHeight: coverHeight, offsetWidth: coverWidth } =
+                getPositionedParent($el) || parent($el);
+            const coverDim = cover(dim, {
+                width: coverWidth + (coverWidth % 2 ? 1 : 0),
+                height: coverHeight + (coverHeight % 2 ? 1 : 0),
+            });
+
+            if (!coverDim.width || !coverDim.height) {
                 return false;
             }
 
-            return dim;
+            return coverDim;
         },
 
         write({ height, width }) {