+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Fix dropbar dropdown alignment if boundary-align is true
 - Fix viewport helper functions take border into account
 - Fix Sticky component responsiveness on window resize
+- Fix Cover component correctly applies ratio if only one of width/height is provided
 
 ## 3.13.10 (April 20, 2022)